+++ Runtime/Debugging/IInjectorDebugger.cs
@@ -1,0 +1,8 @@
+﻿namespace DSystem.Debugging
+{
+    public interface IInjectorDebugger
+    {
+        public void StartInjection(object instance);
+        public void EndInjection(object instance);
+    }
+}