+++ Runtime/Debugging/InjectorDebugger.cs
@@ -1,0 +1,19 @@
+﻿using DSystem.Utils.Tree;
+
+namespace DSystem.Debugging
+{
+    public class InjectorDebugger : IInjectorDebugger
+    {
+        public readonly Tree<object> Tree = new ();
+        
+        public void StartInjection(object instance)
+        {
+            Tree.Push(instance);
+        }
+
+        public void EndInjection(object instance)
+        {
+            Tree.Back();
+        }
+    }
+}